requirements.txt
pytorch-lightning
lightning-flash[image]
mlflow
